refactor (fe): moved grade tables over to the index.tsx
